Remove unused internal; some reformats and warning casts

+++ src/games/dekware/cards/ntlink.c
@@ -368,7 +368,7 @@ static PTEXT ObjectVolatileVariableGet( "hand", "cards", "count of cards in hand
 	}
 	else
 	{
-		*ppLastValue = SegCreateFromInt( CountCards( ph ) );
+		*ppLastValue = SegCreateFromInt( (int)CountCards( ph ) );
 	}
 	return *ppLastValue;
 }
